New structure for project
- Broke out functions relating to variables and expressions to
  separate files
- Broke out unit tests
